refactor updateMTOShipmentAgent template and test
