--- src/constants/MoveHistory/EventTemplates/UpdateMTOShipment/updateMTOShipmentAgent.js
@@ -1,43 +1,0 @@
-import { formatMoveHistoryAgent } from 'utils/formatters';
-import o from 'constants/MoveHistory/UIDisplay/Operations';
-import d from 'constants/MoveHistory/UIDisplay/DetailsTypes';
-import a from 'constants/MoveHistory/Database/Actions';
-import t from 'constants/MoveHistory/Database/Tables';
-import { getMtoShipmentLabel } from 'utils/formatMtoShipment';
-
-export default {
-  action: a.UPDATE,
-  eventName: o.updateMTOShipment,
-  tableName: t.mto_agents,
-  detailsType: d.LABELED,
-  getEventNameDisplay: () => 'Updated shipment',
-  getDetailsLabeledDetails: ({ oldValues, changedValues, context }) => {
-    let newChangedValues = {
-      email: oldValues.email,
-      first_name: oldValues.first_name,
-      last_name: oldValues.last_name,
-      phone: oldValues.phone,
-      ...changedValues,
-    };
-
-    const agent = formatMoveHistoryAgent(newChangedValues);
-
-    const agentType = changedValues.agent_type ?? oldValues.agent_type;
-
-    let agentLabel = '';
-    if (agentType === 'RECEIVING_AGENT') {
-      agentLabel = 'receiving_agent';
-    } else if (agentType === 'RELEASING_AGENT') {
-      agentLabel = 'releasing_agent';
-    }
-
-    newChangedValues = {
-      ...changedValues,
-      ...getMtoShipmentLabel({ context }),
-    };
-
-    newChangedValues[agentLabel] = agent;
-
-    return newChangedValues;
-  },
-};
